+++ src/customComponents/ProjectProfile/hooks/index.tsx
@@ -1,0 +1,8 @@
+import useCid from './hooks';
+import useAverage from './useAverage';
+import useProject from './useProject';
+import useReviews from './useReviews';
+import { useReviewSend } from './useReviewSend';
+import useProjectMeta from './useProjectMeta';
+
+export { useCid, useAverage, useReviewSend, useReviews, useProject, useProjectMeta };
