+++ src/customComponents/ProjectProfile/hooks/useProjectMeta.tsx
@@ -1,7 +1,7 @@
 import { useQuery, UseQueryResult } from 'react-query';
-import { ProjectAl } from '../../interfaces';
-import { NewMetaData } from '../../typeSystem/jsonTypes';
-import { populateMetadata } from './majorUtils';
+import { ProjectAl } from '../../../interfaces';
+import { NewMetaData } from '../../../typeSystem/jsonTypes';
+import { populateMetadata } from '../majorUtils';
 
 export default function useProjectMeta(
   project: ProjectAl,
